Simplify SendForm dataflow
